+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Amber, Lime, Yellow palettes and swatches - [#42](https://github.com/dartoos-dev/eo_color/issues/49).
+  - **Palettes:** Amber, AmberAccent; Lime, LimeAccent; Yellow, YellowAccent.
+  - **Swatches:** Ambers, AmberAccents; Limes, LimeAccents; Yellows,
+    YellowAccents.
+
 ### Changed
 - rearrangement in src directory structure - [#50](https://github.com/dartoos-dev/eo_color/issues/50).